assignment 3

+++ routes/dishRouter.js
@@ -12,7 +12,6 @@ dishRouter.use(bodyParser.json());
 dishRouter.route('/')
 .get((req,res,next) => {
     Dishes.find({})
-    .populate('comments.author')
     .then((dishes) => {
         res.statusCode = 200;
         res.setHeader('Content-Type', 'application/json');
@@ -21,6 +20,7 @@ dishRouter.route('/')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser,(req, res, next) => {
+    authenticate.verifyAdmin(req,res,next) ;
     Dishes.create(req.body)
     .then((dish) => {
         console.log('Dish Created ', dish);
@@ -31,10 +31,12 @@ dishRouter.route('/')
     .catch((err) => next(err));
 })
 .put(authenticate.verifyUser,(req, res, next) => {
+    authenticate.verifyAdmin(req,res,next) ;
     res.statusCode = 403;
     res.end('PUT operation not supported on /dishes');
 })
 .delete(authenticate.verifyUser,(req, res, next) => {
+    authenticate.verifyAdmin(req,res,next) ;
     Dishes.remove({})
     .then((resp) => {
         res.statusCode = 200;
@@ -56,10 +58,12 @@ dishRouter.route('/:dishId')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser,(req, res, next) => {
+    authenticate.verifyAdmin(req,res,next) ;
     res.statusCode = 403;
     res.end('POST operation not supported on /dishes/'+ req.params.dishId);
 })
 .put(authenticate.verifyUser,(req, res, next) => {
+    authenticate.verifyAdmin(req,res,next) ;
     Dishes.findByIdAndUpdate(req.params.dishId, {
         $set: req.body
     }, { new: true })
@@ -71,6 +75,7 @@ dishRouter.route('/:dishId')
     .catch((err) => next(err));
 })
 .delete(authenticate.verifyUser,(req, res, next) => {
+    authenticate.verifyAdmin(req,res,next) ;
     Dishes.findByIdAndRemove(req.params.dishId)
     .then((resp) => {
         res.statusCode = 200;
@@ -103,7 +108,7 @@ dishRouter.route('/:dishId/comments')
     .then((dish) => {
         if (dish != null) {
             req.body.author = req.user._id;
-            dish.comments.push(req.body);
+            dish.comments = dish.comments.concat(req.body);
             dish.save()
             .then((dish) => {
                 res.statusCode = 200;
@@ -125,6 +130,7 @@ dishRouter.route('/:dishId/comments')
         + req.params.dishId + '/comments');
 })
 .delete(authenticate.verifyUser,(req, res, next) => {
+    authenticate.verifyAdmin(req,res,next) ;
     Dishes.findById(req.params.dishId)
     .then((dish) => {
         if (dish != null) {
@@ -177,8 +183,9 @@ dishRouter.route('/:dishId/comments/:commentId')
 })
 .put(authenticate.verifyUser,(req, res, next) => {
     Dishes.findById(req.params.dishId)
+    .populate('comments.author')
     .then((dish) => {
-        if (dish != null && dish.comments.id(req.params.commentId) != null) {
+        if (dish != null && dish.comments.id(req.params.commentId).author._id.equals(req.user._id)) {
             if (req.body.rating) {
                 dish.comments.id(req.params.commentId).rating = req.body.rating;
             }
@@ -197,18 +204,24 @@ dishRouter.route('/:dishId/comments/:commentId')
             err.status = 404;
             return next(err);
         }
-        else {
+        else if(dish.comments.id(req.params.commentId) == null) {
             err = new Error('Comment ' + req.params.commentId + ' not found');
             err.status = 404;
             return next(err);
+        }
+        else {
+          err = new Error("you can't edit comments that you didn't write! \n");
+          err.status = 403 ;
+          return next(err) ;
         }
     }, (err) => next(err))
     .catch((err) => next(err));
 })
 .delete(authenticate.verifyUser,(req, res, next) => {
-    Dishes.findById(req.params.dishId)
+  Dishes.findById(req.params.dishId)
+  .populate('comments.author')
     .then((dish) => {
-        if (dish != null && dish.comments.id(req.params.commentId) != null) {
+        if (dish != null && dish.comments.id(req.params.commentId).author._id.equals(req.user._id)) {
             dish.comments.id(req.params.commentId).remove();
             dish.save()
             .then((dish) => {
@@ -222,10 +235,15 @@ dishRouter.route('/:dishId/comments/:commentId')
             err.status = 404;
             return next(err);
         }
-        else {
+        else if (dish.comments.id(req.params.commentId) == null) {
             err = new Error('Comment ' + req.params.commentId + ' not found');
             err.status = 404;
             return next(err);
+        }
+        else {
+          err = new Error("you can't delete comments that you didn't write! \n");
+          err.status = 403 ;
+          return next(err) ;
         }
     }, (err) => next(err))
     .catch((err) => next(err));
