+++ routes/leaderRouter.js
@@ -20,6 +20,7 @@ leaderRouter.route('/')
   .catch((err)=>{next(err);});
 })
 .post(authenticate.verifyUser,(req, res, next) =>{
+  authenticate.verifyAdmin(req,res,next) ;
   Leaders.create(req.body)
   .then((leader)=>{
     console.log('leader added ' , leader);
@@ -30,10 +31,12 @@ leaderRouter.route('/')
   .catch((err)=>next(err));
 })
 .put(authenticate.verifyUser,(req, res, next) =>{
+  authenticate.verifyAdmin(req,res,next) ;
   res.statusCode = 403 ;
   res.end('the PUT method is not available for /leaders');
 })
 .delete(authenticate.verifyUser,(req, res, next) =>{
+  authenticate.verifyAdmin(req,res,next) ;
   Leaders.remove({})
   .then((resp)=>{
     res.statusCode = 200 ;
@@ -54,10 +57,12 @@ leaderRouter.route('/:leaderId')
   .catch((err)=>next(err));
 })
 .post(authenticate.verifyUser,(req, res, next) =>{
+  authenticate.verifyAdmin(req,res,next) ;
   res.statusCode = 403 ;
   res.end('the POST method is not available for /leaders/'+req.params.leaderId);
 })
 .put(authenticate.verifyUser,(req, res, next) =>{
+  authenticate.verifyAdmin(req,res,next) ;
   Leaders.findByIdAndUpdate(req.params.leaderId,{
     $set : req.body
   },{
@@ -71,6 +76,7 @@ leaderRouter.route('/:leaderId')
   .catch((err)=>next(err));
 })
 .delete(authenticate.verifyUser,(req, res, next) =>{
+  authenticate.verifyAdmin(req,res,next) ;
   Leaders.findByIdAndRemove(req.params.leaderId)
   .then((resp)=>{
     res.statusCode = 200 ;
